requirements.txt
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
joblib==0.16.0
more-itertools==8.5.0
numpy==1.18.5
packaging==20.4
pandas==0.24.2
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.0.2
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
threadpoolctl==2.1.0
toml==0.10.1
urllib3==1.25.10
zipp==3.1.0
